feat(dedicated.netapp): etf refactor network labels
ref: MANAGER-12423
